🧹 Refactor Matrix.tsx to improve maintainability
